Bug fix rabbits repelling each other
When the higher player number is on the right and the bunnies clash in the water, they cancel out each others' motion.
When the lower player number is on the right, one player manages to push the other along. This is unexpected behaviour.
Since the behaviour from the first one looks right to me, I've inverted it to use in the second case.